+++ app/assets/scripts/minireel/campaign.js
@@ -48,8 +48,10 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
 
         .config(['c6StateProvider',
         function( c6StateProvider ) {
-            c6StateProvider.state('MR:Campaigns', ['$injector','paginatedDbList',
-            function                              ( $injector , paginatedDbList ) {
+            c6StateProvider.state('MR:Campaigns', ['$injector','paginatedDbList','c6State',
+            function                              ( $injector , paginatedDbList , c6State ) {
+                var PortalState = c6State.get('Portal');
+
                 $injector.invoke(PaginatedListState, this);
 
                 this.templateUrl = 'views/minireel/campaigns.html';
@@ -61,7 +63,8 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                 };
                 this.model = function() {
                     return paginatedDbList('campaign', {
-                        sort: 'lastUpdated,-1'
+                        sort: 'lastUpdated,-1',
+                        org: PortalState.cModel.org.id
                     }, this.limit, this.page).ensureResolution();
                 };
             }]);
@@ -109,6 +112,20 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                     return minireels.indexOf(minireel) === index;
                 });
             };
+        }])
+
+        .config(['c6StateProvider',
+        function( c6StateProvider ) {
+            c6StateProvider.state('MR:Campaign.Embed', [function() {
+                this.templateUrl = 'views/minireel/campaigns/campaign/embed.html';
+                this.controller = 'CampaignEmbedController';
+                this.controllerAs = 'CampaignEmbedCtrl';
+            }]);
+        }])
+
+        .controller('CampaignEmbedController', ['cState',
+        function(cState) {
+            this.parentState = cState.cParent.cName;
         }])
 
         .config(['c6StateProvider',
@@ -342,8 +359,8 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
             }]);
         }])
 
-        .controller('CampaignMiniReelsController', ['$scope',
-        function                                   ( $scope ) {
+        .controller('CampaignMiniReelsController', ['$scope','$window','MiniReelService',
+        function                                   ( $scope , $window , MiniReelService ) {
             var CampaignCtrl = $scope.CampaignCtrl,
                 campaign = CampaignCtrl.model;
 
@@ -375,6 +392,19 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                     item: minireel
                 }, data));
                 return minireel;
+            };
+
+            this.previewUrlOf = function(minireel) {
+                return MiniReelService.previewUrlOf(minireel);
+            };
+
+            this.previewMiniReel = function(minireel) {
+                var url = MiniReelService.previewUrlOf(minireel),
+                    id = CampaignCtrl.model.id;
+
+                CampaignCtrl.save().then(function() {
+                    $window.open(url + '&campaign=' + id);
+                });
             };
         }])
 
@@ -594,7 +624,8 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                 this.afterModel = function() {
                     this.metaData = {
                         endDate: null,
-                        name: null
+                        name: null,
+                        reportingId: null
                     };
                 };
 
@@ -620,7 +651,8 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
 
                     this.metaData = {
                         endDate: item.endDate,
-                        name: item.name
+                        name: item.name,
+                        reportingId: item.reportingId
                     };
                 };
 
@@ -944,9 +976,12 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
             }]);
         }])
 
-        .controller('CampaignPlacementsController', ['$scope','c6State','$injector','cState',
-        function                                    ( $scope , c6State , $injector , cState ) {
-            var CampaignPlacementsCtrl = this;
+        .controller('CampaignPlacementsController', ['$scope','$injector','$window',
+                                                     'c6State','cState','MiniReelService',
+        function                                    ( $scope , $injector , $window ,
+                                                      c6State , cState , MiniReelService ) {
+            var CampaignPlacementsCtrl = this,
+                CampaignCtrl = $scope.CampaignCtrl;
 
             function overwrite(array, newArray) {
                 array.length = 0;
@@ -996,13 +1031,23 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                 }));
             };
 
-            this.add = function(minireel) {
-                return c6State.goTo('MR:Placements.MiniReel', [
-                    this.model[this.model.push({
-                        minireel: minireel,
-                        cards: createCardEntries(minireel)
-                    }) - 1]
-                ]);
+            this.add = function(placement) {
+                var exists = this.model.filter(function(existing) {
+                    if (placement.minireel.id === existing.minireel.id) {
+                        /* jshint boss:true */
+                        return (existing.cards = placement.cards);
+                    }
+                    return false;
+                }).length;
+
+                return exists ? this.model : this.model.push(placement);
+            };
+
+            this.use = function(minireel) {
+                return c6State.goTo('MR:Placements.MiniReel', [{
+                    minireel: minireel,
+                    cards: createCardEntries(minireel)
+                }]);
             };
 
             this.remove = function(minireel) {
@@ -1025,6 +1070,19 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                 return CampaignPlacementsCtrl.model.map(function(entry) {
                     return entry.minireel;
                 }).indexOf(minireel) < 0;
+            };
+
+            this.previewUrlOf = function(minireel) {
+                return MiniReelService.previewUrlOf(minireel);
+            };
+
+            this.previewMiniReel = function(minireel) {
+                var url = MiniReelService.previewUrlOf(minireel),
+                    id = CampaignCtrl.model.id;
+
+                CampaignCtrl.save().then(function() {
+                    $window.open(url + '&campaign=' + id);
+                });
             };
 
             $scope.$on('CampaignCtrl:campaignDidSave', function() {
@@ -1061,6 +1119,7 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
         .controller('PlacementsMiniReelController', ['$scope','c6State',
         function                                    ( $scope , c6State ) {
             var CampaignCtrl = $scope.CampaignCtrl,
+                CampaignPlacementsCtrl = $scope.CampaignPlacementsCtrl,
                 cards = CampaignCtrl.model.cards;
 
             this.cardOptions = cards.reduce(function(cardOptions, data) {
@@ -1080,10 +1139,18 @@ function( angular , c6State  , PaginatedListState          , PaginatedListContro
                 this.original = entry;
             };
 
+            this.valid = function() {
+                return this.model.cards.some(function(card) {
+                    return !!card.wildcard;
+                });
+            };
+
             this.confirm = function() {
                 extend(this.original, {
                     cards: this.model.cards
                 });
+
+                CampaignPlacementsCtrl.add(this.original);
 
                 return c6State.goTo('MR:Campaign.Placements');
             };